Change project structure, split methods into multiple files.
