+++ src/pmp/python/pyproject.toml
@@ -1,2 +1,2 @@
 [tool.pytest.ini_options]
-pythonpath = ["./src"]
+pythonpath = ["."]
