--- src/pmp/python/src/pypmp/__init__.pyi
@@ -1,1 +1,0 @@
-from .pypmp import *
